+++ app/views/businessdetails/vat_registered.scala.html
