+++ test/views/businessdetails/activity_start_dateSpec.scala
@@ -14,41 +14,39 @@
  * limitations under the License.
  */
 
-package views.aboutthebusiness
+package views.businessdetails
 
 import forms.{EmptyForm, Form2, InvalidForm, ValidForm}
-import models.businessdetails.{VATRegistered, VATRegisteredYes}
+import jto.validation.{Path, ValidationError}
+import models.businessdetails.ActivityStartDate
+import org.joda.time.LocalDate
 import org.scalatest.MustMatchers
-import utils.AmlsSpec
-import jto.validation.Path
-import jto.validation.ValidationError
 import play.api.i18n.Messages
+import utils.AmlsSpec
 import views.Fixture
 
 
-class vat_registeredSpec extends AmlsSpec with MustMatchers  {
-
+class activity_start_dateSpec extends AmlsSpec with MustMatchers {
   trait ViewFixture extends Fixture {
     implicit val requestWithToken = addToken(request)
   }
-
-  "vat_registered view" must {
+  "activity_start_date view" must {
     "have correct title" in new ViewFixture {
 
-      val form2: ValidForm[VATRegistered] = Form2(VATRegisteredYes("1234"))
+      val form2: ValidForm[ActivityStartDate] = Form2(ActivityStartDate(LocalDate.now))
 
-      def view = views.html.aboutthebusiness.vat_registered(form2, true)
+      def view = views.html.aboutthebusiness.activity_start_date(form2, true)
 
-      doc.title must startWith(Messages("businessdetails.registeredforvat.title") + " - " + Messages("summary.aboutbusiness"))
+      doc.title must startWith(Messages("businessdetails.activity.start.date.title") + " - " + Messages("summary.aboutbusiness"))
     }
 
     "have correct headings" in new ViewFixture {
 
-      val form2: ValidForm[VATRegistered] = Form2(VATRegisteredYes("1234"))
+      val form2: ValidForm[ActivityStartDate] = Form2(ActivityStartDate(LocalDate.now))
 
-      def view = views.html.aboutthebusiness.vat_registered(form2, true)
+      def view = views.html.aboutthebusiness.activity_start_date(form2, true)
 
-      heading.html must be(Messages("businessdetails.registeredforvat.title"))
+      heading.html must be(Messages("businessdetails.activity.start.date.title"))
       subHeading.html must include(Messages("summary.aboutbusiness"))
 
     }
@@ -57,27 +55,22 @@ class vat_registeredSpec extends AmlsSpec with MustMatchers  {
 
       val form2: InvalidForm = InvalidForm(Map.empty,
         Seq(
-          (Path \ "registeredForVAT") -> Seq(ValidationError("not a message Key")),
-          (Path \ "vrnNumber-panel") -> Seq(ValidationError("second not a message Key"))
+          (Path \ "startDate") -> Seq(ValidationError("not a message Key"))
         ))
 
-      def view = views.html.aboutthebusiness.vat_registered(form2, true)
+      def view = views.html.aboutthebusiness.activity_start_date(form2, true)
 
       errorSummary.html() must include("not a message Key")
-      errorSummary.html() must include("second not a message Key")
 
-      doc.getElementById("registeredForVAT")
+      doc.getElementById("startDate")
         .getElementsByClass("error-notification").first().html() must include("not a message Key")
-
-      doc.getElementById("vrnNumber-panel")
-        .getElementsByClass("error-notification").first().html() must include("second not a message Key")
 
     }
 
     "have a back link" in new ViewFixture {
       val form2: Form2[_] = EmptyForm
 
-      def view = views.html.aboutthebusiness.vat_registered(form2, true)
+      def view = views.html.aboutthebusiness.activity_start_date(form2, true)
 
       doc.getElementsByAttributeValue("class", "link-back") must not be empty
     }
