+++ test/views/businessdetails/contacting_you_phoneSpec.scala
@@ -14,18 +14,18 @@
  * limitations under the License.
  */
 
-package views.aboutthebusiness
+package views.businessdetails
 
 import forms.{EmptyForm, Form2, InvalidForm, ValidForm}
 import jto.validation.{Path, ValidationError}
-import models.businessdetails.ContactingYouEmail
+import models.businessdetails.ContactingYouPhone
 import org.scalatest.MustMatchers
 import play.api.i18n.Messages
 import utils.AmlsSpec
 import views.Fixture
 
 
-class contacting_youSpec extends AmlsSpec with MustMatchers  {
+class contacting_you_phoneSpec extends AmlsSpec with MustMatchers  {
 
   trait ViewFixture extends Fixture {
     implicit val requestWithToken = addToken(request)
@@ -34,21 +34,20 @@ class contacting_youSpec extends AmlsSpec with MustMatchers  {
   "contacting_you view" must {
     "have correct title, headings and form fields" in new ViewFixture {
 
-      val form2: ValidForm[ContactingYouEmail] = Form2(ContactingYouEmail("123456789789","test@test.com"))
+      val form2: ValidForm[ContactingYouPhone] = Form2(ContactingYouPhone("123456789789"))
 
       def view = {
-        views.html.aboutthebusiness.contacting_you(form2, true)
+        views.html.aboutthebusiness.contacting_you_phone(form2, true)
       }
 
-      doc.title must be(Messages("businessdetails.contactingyou.email.title") +
+      doc.title must be(Messages("businessdetails.contactingyou.phone.title") +
         " - " + Messages("summary.aboutbusiness") +
         " - " + Messages("title.amls") +
         " - " + Messages("title.gov"))
-      heading.html must be(Messages("businessdetails.contactingyou.email.title"))
+      heading.html must be(Messages("businessdetails.contactingyou.phone.title"))
       subHeading.html must include(Messages("summary.aboutbusiness"))
 
-      doc.getElementsByAttributeValue("name", "email") must not be empty
-      doc.getElementsByAttributeValue("name", "confirmEmail") must not be empty
+      doc.getElementsByAttributeValue("name", "phoneNumber") must not be empty
 
     }
 
@@ -60,16 +59,17 @@ class contacting_youSpec extends AmlsSpec with MustMatchers  {
         ))
 
       def view = {
-        views.html.aboutthebusiness.contacting_you(form2, true)
+        views.html.aboutthebusiness.contacting_you_phone(form2, true)
       }
 
       errorSummary.html() must include("not a message Key")
+
     }
 
     "have a back link" in new ViewFixture {
       val form2: Form2[_] = EmptyForm
 
-      def view = views.html.aboutthebusiness.contacting_you(form2, true)
+      def view = views.html.aboutthebusiness.contacting_you_phone(form2, true)
 
       doc.getElementsByAttributeValue("class", "link-back") must not be empty
     }
