+++ test/views/businessdetails/letters_addressSpec.scala
@@ -14,52 +14,54 @@
  * limitations under the License.
  */
 
-package views.aboutthebusiness
+package views.businessdetails
 
 import forms.{EmptyForm, Form2, InvalidForm, ValidForm}
 import jto.validation.{Path, ValidationError}
-import models.businessdetails.ContactingYouPhone
+import models.businessdetails.{ConfirmRegisteredOffice, RegisteredOfficeUK}
 import org.scalatest.MustMatchers
 import play.api.i18n.Messages
 import utils.AmlsSpec
 import views.Fixture
 
 
-class contacting_you_phoneSpec extends AmlsSpec with MustMatchers  {
+class letters_addressSpec extends AmlsSpec with MustMatchers {
 
   trait ViewFixture extends Fixture {
     implicit val requestWithToken = addToken(request)
   }
 
-  "contacting_you view" must {
+  "letters_address view" must {
     "have correct title, headings and form fields" in new ViewFixture {
 
-      val form2: ValidForm[ContactingYouPhone] = Form2(ContactingYouPhone("123456789789"))
+      val form2: ValidForm[ConfirmRegisteredOffice] = Form2(ConfirmRegisteredOffice(true))
 
       def view = {
-        views.html.aboutthebusiness.contacting_you_phone(form2, true)
+        val address = RegisteredOfficeUK("line1","line2",None,None,"AB12CD")
+        views.html.aboutthebusiness.letters_address(form2, address, true)
       }
 
-      doc.title must be(Messages("businessdetails.contactingyou.phone.title") +
+      doc.title must be(Messages("businessdetails.lettersaddress.title") +
         " - " + Messages("summary.aboutbusiness") +
         " - " + Messages("title.amls") +
         " - " + Messages("title.gov"))
-      heading.html must be(Messages("businessdetails.contactingyou.phone.title"))
+      heading.html must be(Messages("businessdetails.lettersaddress.title"))
       subHeading.html must include(Messages("summary.aboutbusiness"))
 
-      doc.getElementsByAttributeValue("name", "phoneNumber") must not be empty
-
+      doc.getElementsMatchingOwnText("line1").text mustBe "line1 line2 AB12CD"
+      doc.select("input[type=radio]").size mustBe 2
     }
 
     "show error summary in correct location" in new ViewFixture {
 
       val form2: InvalidForm = InvalidForm(Map.empty,
         Seq(
-          (Path \ "contactingyou-fieldset") -> Seq(ValidationError("not a message Key"))
+          (Path \ "lettersAddress") -> Seq(ValidationError("not a message Key"))
         ))
 
       def view = {
-        views.html.aboutthebusiness.contacting_you_phone(form2, true)
+        val address = RegisteredOfficeUK("line1","line2",None,None,"AB12CD")
+        views.html.aboutthebusiness.letters_address(form2, address, true)
       }
 
       errorSummary.html() must include("not a message Key")
@@ -69,7 +71,10 @@ class contacting_you_phoneSpec extends AmlsSpec with MustMatchers  {
     "have a back link" in new ViewFixture {
       val form2: Form2[_] = EmptyForm
 
-      def view = views.html.aboutthebusiness.contacting_you_phone(form2, true)
+      def view = {
+        val address = RegisteredOfficeUK("line1","line2",None,None,"AB12CD")
+        views.html.aboutthebusiness.letters_address(form2, address, true)
+      }
 
       doc.getElementsByAttributeValue("class", "link-back") must not be empty
     }
